Face Mask Detection System mini project
This mini project named as 'Face Mask Detection System' was made using Python on Google Colab. Here, I've used 2 of my own images, one with a mask and the other without a mask, to build a CNN model using TensorFlow to detect if which image has a person wearing a mask and which doesn't.